Revert "Undo gRPC changes"
This reverts commit 4c0ac4c6c458ae0c265fce496741c7b759f749b3.
