+++ beacon-chain/rpc/slasher/server.go
@@ -2,5 +2,11 @@
 // which allows for checking if attestations or blocks are slashable.
 package slasher
 
+import (
+	"github.com/prysmaticlabs/prysm/beacon-chain/slasher"
+)
+
 // Server defines a server implementation of the gRPC slasher service.
-type Server struct{}
+type Server struct {
+	slasher slasher.SlashingChecker
+}
