+++ beacon-chain/slasher/detect_blocks_test.go
@@ -12,7 +12,6 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/event"
 	"github.com/prysmaticlabs/prysm/shared/params"
-	"github.com/prysmaticlabs/prysm/shared/testutil/assert"
 	"github.com/prysmaticlabs/prysm/shared/testutil/require"
 	logTest "github.com/sirupsen/logrus/hooks/test"
 )
@@ -47,62 +46,6 @@ func Test_processQueuedBlocks_DetectsDoubleProposals(t *testing.T) {
 	cancel()
 	<-exitChan
 	require.LogsContain(t, hook, "Proposer slashing detected")
-}
-
-func TestIsSlashableBlock(t *testing.T) {
-	ctx := context.Background()
-	beaconDB := dbtest.SetupDB(t)
-	s := &Service{
-		serviceCfg: &ServiceConfig{
-			Database:              beaconDB,
-			ProposerSlashingsFeed: new(event.Feed),
-		},
-		params:    DefaultParams(),
-		blksQueue: newBlocksQueue(),
-	}
-	err := beaconDB.SaveBlockProposals(ctx, []*slashertypes.SignedBlockHeaderWrapper{
-		createProposalWrapper(t, 2, 3, []byte{1}),
-		createProposalWrapper(t, 3, 3, []byte{1}),
-	})
-	require.NoError(t, err)
-	tests := []struct {
-		name              string
-		blockToCheck      *slashertypes.SignedBlockHeaderWrapper
-		shouldBeSlashable bool
-	}{
-		{
-			name:              "should not detect if same signing root",
-			blockToCheck:      createProposalWrapper(t, 2, 3, []byte{1}),
-			shouldBeSlashable: false,
-		},
-		{
-			name:              "should not detect if different slot",
-			blockToCheck:      createProposalWrapper(t, 1, 3, []byte{2}),
-			shouldBeSlashable: false,
-		},
-		{
-			name:              "should not detect if different validator index",
-			blockToCheck:      createProposalWrapper(t, 2, 4, []byte{2}),
-			shouldBeSlashable: false,
-		},
-		{
-			name:              "detects differing signing root",
-			blockToCheck:      createProposalWrapper(t, 2, 3, []byte{2}),
-			shouldBeSlashable: true,
-		},
-		{
-			name:              "should detect another slot",
-			blockToCheck:      createProposalWrapper(t, 3, 3, []byte{2}),
-			shouldBeSlashable: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			doubleProposal, err := s.IsSlashableBlock(ctx, tt.blockToCheck.SignedBeaconBlockHeader)
-			require.NoError(t, err)
-			assert.Equal(t, tt.shouldBeSlashable, doubleProposal != nil)
-		})
-	}
 }
 
 func Test_processQueuedBlocks_NotSlashable(t *testing.T) {
